requirements.txt
asgiref==3.2.3
dj-database-url==0.5.0
Django==3.0.1
django-cors-headers==3.2.0
django-heroku==0.3.1
djangorestframework==3.11.0
gunicorn==20.0.4
Pillow==6.2.1
psycopg2==2.8.4
pytz==2019.3
sqlparse==0.3.0
whitenoise==5.0.1
